UHF-11725: Add tests
