+++ public/modules/custom/grants_handler/tests/src/Kernel/Block/ServicePageAnonBlockCacheExpirationTest.php
@@ -4,7 +4,7 @@ namespace Drupal\Tests\grants_handler\Kernel\Block;
 
 use Drupal\grants_handler\ApplicationStatusServiceInterface;
 use Drupal\grants_handler\ServicePageBlockService;
-use Drupal\KernelTests\KernelTestBase;
+use Drupal\Tests\grants_handler\Kernel\GrantsHandlerKernelTestBase;
 use Drupal\webform\Entity\Webform;
 use Drupal\grants_handler\Plugin\Block\ServicePageAnonBlock;
 use Drupal\Tests\node\Traits\NodeCreationTrait;
@@ -15,90 +15,10 @@ use Prophecy\PhpUnit\ProphecyTrait;
  *
  * @group grants_handler
  */
-class ServicePageAnonBlockCacheExpirationTest extends KernelTestBase {
+class ServicePageAnonBlockCacheExpirationTest extends GrantsHandlerKernelTestBase {
 
   use NodeCreationTrait;
   use ProphecyTrait;
-
-  /**
-   * {@inheritdoc}
-   */
-  protected static $modules = [
-    'content_translation',
-    'externalauth',
-    'entity_reference_revisions',
-    'grants_handler',
-    'grants_profile',
-    'grants_mandate',
-    'grants_metadata',
-    'grants_attachments',
-    'grants_events',
-    'helfi_yjdh',
-    'helfi_audit_log',
-    'locale',
-    'language',
-    'block',
-    'path_alias',
-    'file',
-    'field',
-    'helfi_api_base',
-    'helfi_atv',
-    'helfi_helsinki_profiili',
-    'helfi_tunnistamo',
-    'node',
-    'openid_connect',
-    'options',
-    'openid_connect_logout_redirect',
-    'paragraphs',
-    'system',
-    'taxonomy',
-    'text',
-    'user',
-    'webform',
-  ];
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function setUp(): void {
-    parent::setUp();
-    $this->installSchema('webform', ['webform']);
-    $this->installSchema('locale', [
-      'locales_source',
-      'locales_target',
-      'locales_location',
-    ]);
-    $this->installEntitySchema('webform');
-    $this->installEntitySchema('node');
-    $this->installEntitySchema('user');
-    $this->installEntitySchema('taxonomy_term');
-    $this->installEntitySchema('paragraph');
-    $this->installEntitySchema('paragraphs_type');
-
-    $this->installConfig([
-      'externalauth',
-      'grants_profile',
-      'grants_mandate',
-      'grants_metadata',
-      'grants_attachments',
-      'grants_events',
-      'grants_handler',
-      'helfi_yjdh',
-      'helfi_audit_log',
-      'locale',
-      'language',
-      'file',
-      'field',
-      'helfi_api_base',
-      'helfi_atv',
-      'helfi_tunnistamo',
-      'openid_connect',
-      'openid_connect_logout_redirect',
-      'paragraphs',
-      'system',
-      'webform',
-    ]);
-  }
 
   /**
    * Test cache max age changes when webform start/end date changes.
